agentcore/deployment/agent/requirements.txt
# Core AWS and deployment dependencies
boto3>=1.26.0
botocore>=1.29.0
pyyaml>=6.0
requests>=2.28.0
aws-opentelemetry-distro>=0.10.0

# AgentCore dependencies (without starter toolkit to avoid CodeBuild issues)
bedrock-agentcore>=0.1.0
strands-agents
strands-agents-tools
strands-agents[a2a]>=1.3.0

a2a-sdk[sql]>=0.3.0
# Utility dependencies
packaging>=21.0
setuptools>=65.0
wheel>=0.37.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0

# Optional dependencies
IPython
uv

# Knowledge MCP Server Dependencies
fastmcp
mcp
starlette
uvicorn

gql
requests_aws4auth
requests_toolbelt
requests_aws_sign
websocket-client
pydantic
typing
